Makefile: WARNINGS: add -Wdouble-promotion
-Wdouble-promotion (C, C++, Objective-C and Objective-C++ only)
   Give a warning when a value of type "float" is implicitly promoted to
   "double".  CPUs with a 32-bit "single-precision" floating-point unit
   implement "float" in hardware, but emulate "double" in software.  On
   such a machine, doing computations using "double" values is much more
   expensive because of the overhead required for software emulation.

   It is easy to accidentally do computations with "double" because
   floating-point literals are implicitly of type "double".
   For example, in:

           float area(float radius)
           {
              return 3.14159 * radius * radius;
           }

   the compiler will perform the entire computation with "double"
   because the floating-point literal is a "double".

+++ lib/bmf/src/NetworkInterfaces.c
@@ -804,7 +804,7 @@ void FindNeighbors(
         "%s: ----> forwarding pkt to %s will cost ETX %5.2f\n",
         PLUGIN_NAME_SHORT,
         olsr_ip_to_string(&buf, &walker->neighbor_iface_addr),
-        currEtx);
+        (double)currEtx);
 
       /* If the candidate neighbor is best reached via another interface, then skip 
        * the candidate neighbor; the candidate neighbor has been / will be selected via that
